pyproject.toml
[project]
name = "javansanat"
version = "0.1.0"
description = "a project about the youth industry"
authors = [
    {name = "A.L",email = "abolfazlloukzade7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg (>=3.2.10,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-mptt (>=0.18.0,<0.19.0)",
    "django-resized (>=1.0.3,<2.0.0)",
    "django-ckeditor-5 (>=0.2.18,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
